requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
feedparser
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv